requirements.txt
python-whois
dnspython
requests
os
sys
dns.resolver
sockets
argparse